// src/generated/models/geographyType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * One Health Lyme Gap Atlas API
 * Public API for Atlas data and reviewed knowledge-graph evidence chat.
 * OpenAPI spec version: 0.1.0
 */

export type GeographyType = typeof GeographyType[keyof typeof GeographyType];


export const GeographyType = {
  county: 'county',
  state: 'state',
} as const;